fix: :bug: Turn `pending_submit_proof_requests` into a `BTreeSet` to avoid repetitions and prioritise the latest insertion
