+++ client/blockchain-service/src/types.rs
@@ -1,7 +1,98 @@
+use std::cmp::Ordering;
+
+use codec::{Decode, Encode};
 use frame_support::dispatch::DispatchInfo;
 use frame_system::EventRecord;
+use shc_common::types::{BlockNumber, ProviderId, RandomnessOutput, TrieRemoveMutation};
 use sp_core::H256;
 use sp_runtime::DispatchError;
+
+use crate::events::NewChallengeSeed;
+
+/// A struct that holds the information to submit a storage proof.
+///
+/// This struct is used as an item in the `pending_submit_proof_requests` queue.
+#[derive(Debug, Clone, Encode, Decode)]
+pub struct SubmitProofRequest {
+    pub provider_id: ProviderId,
+    pub tick: BlockNumber,
+    pub seed: RandomnessOutput,
+    pub forest_challenges: Vec<H256>,
+    pub checkpoint_challenges: Vec<(H256, Option<TrieRemoveMutation>)>,
+}
+
+impl SubmitProofRequest {
+    pub fn new(
+        provider_id: ProviderId,
+        tick: BlockNumber,
+        seed: RandomnessOutput,
+        forest_challenges: Vec<H256>,
+        checkpoint_challenges: Vec<(H256, Option<TrieRemoveMutation>)>,
+    ) -> Self {
+        Self {
+            provider_id,
+            tick,
+            seed,
+            forest_challenges,
+            checkpoint_challenges,
+        }
+    }
+
+    pub fn new_from_event(
+        new_challenge_seed_event: NewChallengeSeed,
+        forest_challenges: Vec<H256>,
+        checkpoint_challenges: Vec<(H256, Option<TrieRemoveMutation>)>,
+    ) -> Self {
+        Self {
+            provider_id: new_challenge_seed_event.provider_id,
+            tick: new_challenge_seed_event.tick,
+            seed: new_challenge_seed_event.seed,
+            forest_challenges,
+            checkpoint_challenges,
+        }
+    }
+}
+
+impl Ord for SubmitProofRequest {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.tick.cmp(&other.tick)
+    }
+}
+
+impl PartialOrd for SubmitProofRequest {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+// Two `SubmitProofRequest`s are considered equal if they have the same `tick` and `provider_id`.
+// This helps to identify and remove duplicate requests from the queue.
+impl PartialEq for SubmitProofRequest {
+    fn eq(&self, other: &Self) -> bool {
+        self.tick == other.tick && self.provider_id == other.provider_id
+    }
+}
+
+impl Eq for SubmitProofRequest {}
+
+#[derive(Debug, Clone, Encode, Decode)]
+pub struct ConfirmStoringRequest {
+    pub file_key: H256,
+    pub try_count: u32,
+}
+
+impl ConfirmStoringRequest {
+    pub fn new(file_key: H256) -> Self {
+        Self {
+            file_key,
+            try_count: 0,
+        }
+    }
+
+    pub fn increment_try_count(&mut self) {
+        self.try_count += 1;
+    }
+}
 
 /// Type alias for the events vector.
 ///
